tevye_gpt_server/requirements.txt
structlog==25.4.0
fastapi==0.116.1
uvicorn==0.35.0
SQLAlchemy==2.0.43
python-jose==3.5.0
cryptography==46.0.1
bcrypt==5.0.0
passlib==1.7.4
pydantic-settings==2.11.0
psycopg2-binary==2.9.10
email-validator==2.3.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
aiohttp==3.13.2